+++ FromLocalsToLocals/Views/Home/NewsFeed.cshtml
@@ -1,19 +1,19 @@
 ﻿@model FromLocalsToLocals.Models.ViewModels.PostVM
+@inject SignInManager<AppUser> signinManager
+<div class="row">
 
 
-<ul class="nav nav-tabs">
-    <li role="presentation" class="@(Model.ActiveTab == FromLocalsToLocals.Models.ViewModels.Tab.All? "active" : string.Empty)"><a asp-route-tabname="All" asp-controller="Home" asp-action="SwitchTabs">All</a></li>
-    <li role="presentation" class="@(Model.ActiveTab == FromLocalsToLocals.Models.ViewModels.Tab.Love? "active" : string.Empty)"><a asp-route-tabname="Love" asp-controller="Home" asp-action="SwitchTabs">Love</a></li>
-</ul>
+    <nav class="navbar navbar-expand-lg navbar-light bg-white w-100" style="color: gray; border-radius: 4px;">
+        <button class="navbar-toggler" type="button" data-toggle="collapse" data-target="#navbarNav" aria-controls="navbarNav" aria-expanded="false" aria-label="Toggle navigation">
+            <span class="navbar-toggler-icon"></span>
+        </button>
+        <div class="collapse navbar-collapse myNav" id="navbarNav">
+            @if(signinManager.IsSignedIn(User)){
+            <a class="text-center" style="font-size: 19px;width:50%;" asp-route-tabname="MyFeed" asp-controller="Home" asp-action="SwitchTabs">My Feed</a>
+            <a class="text-center" style="font-size: 19px;width:50%;" asp-route-tabname="AllFeed" asp-controller="Home" asp-action="SwitchTabs">All Feed</a>
+            }
+        </div>
+    </nav>
+</div>
 
-@switch (Model.ActiveTab)
-{
-    case FromLocalsToLocals.Models.ViewModels.Tab.All:
-        @await Component.InvokeAsync("AllNews",Model);
-        break;
-    case FromLocalsToLocals.Models.ViewModels.Tab.Love:
-        @await Component.InvokeAsync("LovedVendorsNews",Model);
-        break;
-    default:
-        break;
-}
+@await Component.InvokeAsync("NewsFeed", Model);
