--- FromLocalsToLocals/Views/Shared/Components/LovedVendorsNews/Default.cshtml
@@ -1,3 +1,0 @@
-﻿
-
-SVEIKI
