+++ FromLocalsToLocals/Views/Vendors/Details.cshtml
@@ -60,7 +60,7 @@
                         </li>
                         <li class="nav-item text-center" style="width: 50%;">
                             <a class="nav-link" style="font-size: 19px;" asp-route-id="Vendor/@Model.ID" asp-controller="Reviews" asp-action="Reviews">@_localizer["Reviews"]</a>
-</li>
+                        </li>
                     </ul>
                 </div>
             </nav>
